Issue #268 - provide mvlc configuration tools
Markers are 32 bits so don't casts the -value down to 16 bits.

+++ main/mvlc/CMarker.cpp
@@ -90,7 +90,7 @@ CMarker::addReadoutList(CVMUSBReadoutList& list)
 {
   unsigned int value = getIntegerParameter("-value");
 
-  list.addMarker(static_cast<uint16_t>(value));
+  list.addMarker(value);
 }
